+++ lib/src/wrapper.dart
@@ -1,0 +1,19 @@
+part of binary_types;
+
+/*
+class BinaryWrapper {
+  final BinaryData _data;
+
+  BinaryWrapper(this._data) {
+    if (_data == null) {
+      throw new ArgumentError.notNull("data");
+    }
+  }
+
+  dynamic operator [](index) => _data[index].value;
+
+  void operator []=(index, value) {
+    _data[index].value = value;
+  }
+}
+*/